+++ homeassistant/components/tellduslive/__init__.py
@@ -4,25 +4,23 @@ Support for Telldus Live.
 For more details about this component, please refer to the documentation at
 https://home-assistant.io/components/tellduslive/
 """
-from datetime import datetime, timedelta
+from datetime import timedelta
 import logging
 
 import voluptuous as vol
 
 from homeassistant.components.discovery import SERVICE_TELLDUSLIVE
 from homeassistant.const import (
-    ATTR_BATTERY_LEVEL, CONF_HOST, CONF_TOKEN, DEVICE_DEFAULT_NAME,
-    EVENT_HOMEASSISTANT_START)
+    CONF_HOST, CONF_TOKEN, EVENT_HOMEASSISTANT_START)
 from homeassistant.helpers import discovery
 import homeassistant.helpers.config_validation as cv
-from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.event import track_point_in_utc_time
 from homeassistant.util.dt import utcnow
 from homeassistant.util.json import load_json, save_json
 
-APPLICATION_NAME = 'Home Assistant'
+from .const import DOMAIN
 
-DOMAIN = 'tellduslive'
+APPLICATION_NAME = 'Home Assistant'
 
 REQUIREMENTS = ['tellduslive==0.10.4']
 
@@ -48,8 +46,6 @@ CONFIG_SCHEMA = vol.Schema({
     }),
 }, extra=vol.ALLOW_EXTRA)
 
-
-ATTR_LAST_UPDATED = 'time_last_updated'
 
 CONFIG_INSTRUCTIONS = """
 To link your TelldusLive account:
@@ -276,91 +272,3 @@ class TelldusLiveClient:
     def is_available(self, device_id):
         """Return device availability."""
         return device_id in self._client.device_ids
-
-
-class TelldusLiveEntity(Entity):
-    """Base class for all Telldus Live entities."""
-
-    def __init__(self, hass, device_id):
-        """Initialize the entity."""
-        self._id = device_id
-        self._client = hass.data[DOMAIN]
-        self._client.entities.append(self)
-        self._name = self.device.name
-        _LOGGER.debug('Created device %s', self)
-
-    def changed(self):
-        """Return the property of the device might have changed."""
-        if self.device.name:
-            self._name = self.device.name
-        self.schedule_update_ha_state()
-
-    @property
-    def device_id(self):
-        """Return the id of the device."""
-        return self._id
-
-    @property
-    def device(self):
-        """Return the representation of the device."""
-        return self._client.device(self.device_id)
-
-    @property
-    def _state(self):
-        """Return the state of the device."""
-        return self.device.state
-
-    @property
-    def should_poll(self):
-        """Return the polling state."""
-        return False
-
-    @property
-    def assumed_state(self):
-        """Return true if unable to access real state of entity."""
-        return True
-
-    @property
-    def name(self):
-        """Return name of device."""
-        return self._name or DEVICE_DEFAULT_NAME
-
-    @property
-    def available(self):
-        """Return true if device is not offline."""
-        return self._client.is_available(self.device_id)
-
-    @property
-    def device_state_attributes(self):
-        """Return the state attributes."""
-        attrs = {}
-        if self._battery_level:
-            attrs[ATTR_BATTERY_LEVEL] = self._battery_level
-        if self._last_updated:
-            attrs[ATTR_LAST_UPDATED] = self._last_updated
-        return attrs
-
-    @property
-    def _battery_level(self):
-        """Return the battery level of a device."""
-        from tellduslive import (BATTERY_LOW,
-                                 BATTERY_UNKNOWN,
-                                 BATTERY_OK)
-        if self.device.battery == BATTERY_LOW:
-            return 1
-        if self.device.battery == BATTERY_UNKNOWN:
-            return None
-        if self.device.battery == BATTERY_OK:
-            return 100
-        return self.device.battery  # Percentage
-
-    @property
-    def _last_updated(self):
-        """Return the last update of a device."""
-        return str(datetime.fromtimestamp(self.device.lastUpdated)) \
-            if self.device.lastUpdated else None
-
-    @property
-    def unique_id(self) -> str:
-        """Return a unique ID."""
-        return self._id
