+++ homeassistant/components/tellduslive/const.py
@@ -1,0 +1,3 @@
+"""Consts used by TelldusLive."""
+
+DOMAIN = 'tellduslive'
